requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain
langchain-community
langchain-postgres
faiss-cpu
langchain-openai
pypdf